rename JOB database table to TICKET
