+++ application/models/ticket.js
@@ -3,8 +3,13 @@ const Schema = mongoose.Schema;
 
 const schema = new Schema({
     products: {
-        type: [{ type : Schema.Types.ObjectId, ref: 'Product' }]
-    }
+        type: [
+            { 
+                type : Schema.Types.ObjectId, 
+                ref: 'Product' 
+            }
+        ]
+    },
 }, { timestamps: true });
 
 const Job = mongoose.model('Job', schema);
