+++ application/views/uploadTicket.ejs
@@ -4,7 +4,7 @@
     <%- include ("flashMessages") %>
 </div>
 
-<form action="/jobs/upload" method="POST" enctype="multipart/form-data">
+<form action="/tickets/upload" method="POST" enctype="multipart/form-data">
     <input type="file" name="job-xml"><br>
     <input type="submit">
 </form>
